Feat. add useGetAllDatabases
